+++ src/components/CodeCell/index.ts
@@ -1,0 +1,3 @@
+import CodeCell from './CodeCell'
+
+export default CodeCell
