+++ src/components/Preview/index.ts
@@ -1,0 +1,3 @@
+import Preview from './Preview'
+
+export default Preview
