--- src/components/common/Form/index.ts
@@ -1,3 +1,0 @@
-import Form from './Form'
-
-export default Form
